feat(api): update catalog route to use static catalogs
The `catalogs` route in the API has been updated to use the `static-catalogs` endpoint instead of the `get-catalogs` endpoint. This change improves the efficiency and performance of retrieving catalogs.

Note: This message assumes that the commit is related to the API functionality and follows the convention of prefixing the commit message with the type of change (e.g., feat, fix, chore).